fixed settings import path and moved logout button to sidebar
