+++ src/components/dashboard/Dashboard.js
@@ -8,14 +8,6 @@ const Dashboard = props => {
   return (
     <section className="ui container">
       <h1>Dashboard</h1>
-      <Button
-        onClick={() => {
-          localStorage.removeItem("token");
-          props.history.push("/signin");
-        }}
-      >
-        Logout
-      </Button>
       <TodoList />
     </section>
   );
